+++ src/actions/parserActions.js
@@ -1,8 +1,6 @@
 export const PARSE = 'parser/PARSE'
-export const SET_STANDARD = 'parser/SET_STANDARD'
 export const ADD_WARRIOR = 'parser/ADD_WARRIOR'
 export const REMOVE_WARRIOR = 'parser/REMOVE_WARRIOR'
-export const RESET_CORE = 'parser/RESET_CORE'
 
 export const PARSE_REQUESTED = 'parser/PARSE_REQUESTED'
 export const ADD_WARRIOR_REQUESTED = 'parser/ADD_WARRIOR_REQUESTED'
@@ -14,4 +12,4 @@ function action(type, payload = {}) {
 
 export const parse = redcode => action(PARSE_REQUESTED, { redcode })
 export const addWarrior = () => action(ADD_WARRIOR_REQUESTED)
-export const removeWarrior = () => action(REMOVE_WARRIOR_REQUESTED)
+export const removeWarrior = index => action(REMOVE_WARRIOR_REQUESTED, { index })
